+++ rabbitmq/CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changes
 
 * [FEATURE] Add a metric about the number of connections to rabbitmq.
+* [FEATURE] Add a metric about the number of each connection states.
 
 1.1.0 / 06-05-2017
 ==================